Refactor query logic and add support for HAVING clause.
