+++ src/QueryBasic.php
@@ -21,6 +21,8 @@ class QueryBasic implements QueryBuilderInterface
     protected DbDriverInterface|null $dbDriver = null;
     protected ?Recursive $recursive = null;
     protected bool $distinct = false;
+    protected array $groupBy = [];
+    protected array $having = [];
 
     public static function getInstance(): QueryBasic
     {
@@ -185,6 +187,33 @@ class QueryBasic implements QueryBuilderInterface
     }
 
     /**
+     * Example:
+     *    $query->groupBy(['name']);
+     *
+     * @param array $fields
+     * @return $this
+     */
+    public function groupBy(array $fields): static
+    {
+        $this->groupBy = array_merge($this->groupBy, $fields);
+
+        return $this;
+    }
+
+    /**
+     * Example:
+     *    $query->having('count(price) > 10');
+     *
+     * @param string $filter
+     * @return $this
+     */
+    public function having(string $filter): static
+    {
+        $this->having[] = $filter;
+        return $this;
+    }
+
+    /**
      * @throws InvalidArgumentException
      */
     protected function getFields(): array
@@ -249,7 +278,23 @@ class QueryBasic implements QueryBuilderInterface
             $params = $subQuery->getParams();
         }
         return [ $table . (!empty($alias) && $table != $alias ? " as " . $alias : ""), $params ];
-    }   
+    }
+
+    protected function addGroupBy(): string
+    {
+        if (empty($this->groupBy)) {
+            return "";
+        }
+        return ' GROUP BY ' . implode(', ', $this->groupBy);
+    }
+
+    protected function addHaving(): string
+    {
+        if (empty($this->having)) {
+            return "";
+        }
+        return ' HAVING ' . implode(' AND ', $this->having);
+    }
 
     /**
      * @param DbDriverInterface|null $dbDriver
@@ -281,6 +326,10 @@ class QueryBasic implements QueryBuilderInterface
             $sql .= ' WHERE ' . $whereStr[0];
             $params = array_merge($params, $whereStr[1]);
         }
+
+        $sql .= $this->addGroupBy();
+
+        $sql .= $this->addHaving();
 
         $sql = ORMHelper::processLiteral($sql, $params);
